wip[array]: execute::Primitive
